test(app-core): isolate plugin registry deadline helper
